fix(v2): untie output head by default — fixes single-step output collapse
Pretraining the large model collapsed reproducibly: after a gradient spike,
train loss jumped to ln(V)=7.6 (uniform output) in ONE step and parked at a
dead fixed point. Root cause (multi-lens review + confirmed by ablation): the
TIED, un-regularized output head. `tie_embeddings=True` made one
`embed_tokens[V,d]` serve as both the input lookup and the output projection
(`logits = x @ embed_tokens.T`) with no logit scale / softcap / z-loss. The CE
gradient `(softmax-onehot)*x` has a column-common component that drives all V
rows toward collinearity -> embed_tokens.T near rank-1 -> constant logits ->
uniform. Global-norm clipping bounds the step's magnitude but not its
DIRECTION, so it doesn't prevent it; larger batch only raised the lr ceiling.
v1 was stable because input embeddings and lm_head were SEPARATE tensors.

Fix:
- ModelConfig.tie_embeddings default True->False. Untied is canonical (model /
  checkpoint / sliced() already supported it; sliced narrows lm_head[:dv,:]).
- Output logits always computed in fp32, even under bf16 AMP (model.py): the
  head matmul upcasts so a bf16 column can't overflow to inf -> softmax NaN ->
  poisoned head. ~1% FLOPs.
- AdamW/Lion mu_dtype bf16->fp32 (trainer.py): a bf16 first moment zeroed small
  surviving components post-spike, biasing the update direction.

Validation: untied large, bs64/eff64, lr ramped to 3e-4 (where tied collapsed
at lr 1.1e-4) survived 6k steps with grad_norm ~0.4 (tied: 4-6, spikes to 166)
and reached loss 3.10. Full suite green; the adapter frozen-head test now
verifies the untied lm_head stays frozen.

+++ pawn/trainer.py
@@ -883,21 +883,22 @@ def make_optimizer(
         inner = optax.adamw(
             learning_rate=lr_schedule,
             weight_decay=cfg.weight_decay,
-            # First moment (``mu``) in bf16 — saves ~half the optimizer
-            # state's HBM footprint (~70 MB at BASE / ~140 MB at LARGE)
-            # and the matching read/write bandwidth on every step. The
-            # second moment (``nu``) stays fp32 because Adam's ``rsqrt``
-            # is sensitive to denominator precision near zero. Standard
-            # practice in optax / Maxtext / Mesh-Transformer-JAX.
-            mu_dtype=jnp.bfloat16,
+            # First moment (``mu``) in fp32. A bf16 ``mu`` (~8-bit mantissa)
+            # rounds the small surviving gradient components to zero on the
+            # step *after* a gradient spike — while the fp32 second moment
+            # (``nu``) still carries the spike's large magnitude — biasing the
+            # (clipped, unit-norm) update toward a degenerate direction right
+            # when the model is most fragile. The ~140 MB HBM saving at LARGE
+            # is not worth that pretraining-stability risk.
+            mu_dtype=jnp.float32,
         )
     elif name == "lion":
         inner = optax.lion(
             learning_rate=lr_schedule,
             weight_decay=cfg.weight_decay,
-            # Lion stores only one moment (the interpolated direction);
-            # keep it bf16 for the same bandwidth win as AdamW.mu.
-            mu_dtype=jnp.bfloat16,
+            # Lion's single moment in fp32 for the same post-spike
+            # direction-stability reason as AdamW.mu above.
+            mu_dtype=jnp.float32,
         )
     else:
         raise ValueError(f"unknown optimizer {name!r}; expected adamw/lion")
